requirements.txt
streamlit==1.29.0
python-docx==0.8.11
pdfkit==1.0.0
python-dateutil==2.8.2
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
google-generativeai
PyPDF2>=3.0.0
spacy>=3.7.2
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0